requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
backtesting==0.6.5
bcrypt==4.3.0
beautifulsoup4==4.13.5
binance-connector==3.12.0
bitarray==3.8.0
bokeh==3.8.0
build==1.3.0
cachetools==6.2.1
ccxt==4.5.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
ckzg==2.1.5
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.1
curl_cffi==0.13.0
cycler==0.12.1
Cython==3.2.0
cytoolz==1.1.0
dateparser==1.2.2
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.116.1
filelock==3.19.1
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.33.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
hyperliquid-python-sdk==0.20.1
idna==3.10
imbalanced-learn==0.14.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonalias==0.1.1
kiwisolver==1.4.9
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
more-itertools==10.8.0
MouseInfo==0.1.3
mplfinance==0.12.10b0
msgpack==1.1.2
mss==10.1.0
multidict==6.7.0
multitasking==0.0.12
narwhals==2.10.2
Nasdaq-Data-Link==1.0.4
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
openai==1.109.0
packaging==25.0
pandas==2.3.2
pandas-ta==0.4.71b0
parsimonious==0.10.0
patsy==1.0.2
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.0
praw==7.8.1
prawcore==2.4.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.5
PyPDF2==3.0.1
pyperclip==1.11.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rlp==4.1.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
solders==0.27.0
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
statsmodels==0.14.5
ta==0.11.0
TA-Lib==0.6.7
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
toolz==1.1.0
tornado==6.5.2
tqdm==4.67.1
tweepy==4.16.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
web3==7.14.0
websocket-client==1.9.0
websockets==15.0.1
win32_setctime==1.2.0
xyzservices==2025.10.0
yarl==1.22.0
yfinance==0.2.66
